ASCII-Art/requirements.txt
pillow==1.1.6
moviepy
numpy
matplotlib
opencv
